+++ crates/vm/levm/src/constants.rs
@@ -29,7 +29,7 @@ pub const SYS_CALL_GAS_LIMIT: u64 = 30000000;
 pub const TX_BASE_COST: u64 = 21000;
 
 // https://eips.ethereum.org/EIPS/eip-7825
-pub const POST_OSAKA_GAS_LIMIT_CAP: u64 = 16777216;
+pub use ethrex_common::constants::POST_OSAKA_GAS_LIMIT_CAP;
 
 pub const MAX_CODE_SIZE: u64 = 0x6000;
 pub const INIT_CODE_MAX_SIZE: usize = 49152;
